Adds MiqServerWorkerManagementBase
Moves "srsly everything" in app/models/miq_server/worker_management.rb
into it's own module for reuse.

Used a concern here since it seems to handle things better, but can't
use `include_concern` because that doesn't support non-standard paths
(yet...)
